loading screen enhancement - move the source code, reduce the size of the index.js, add dependency of the anime.s
