+++ package.json
@@ -35,6 +35,7 @@
     "webpack-dev-server": "^3.8.0"
   },
   "dependencies": {
+    "animejs": "^3.1.0",
     "babel-preset-es2015": "^6.24.1",
     "bootstrap": "^4.3.1",
     "jquery": "^3.4.1",
