+++ src/index.js
@@ -1,56 +1,28 @@
-import React, { lazy, Suspense } from 'react'
-import ReactDOM from 'react-dom'
-import { BrowserRouter } from 'react-router-dom'
-import App from './app.js'
-import './scss/app.scss'
-import './style.less'
+import anime from 'animejs/lib/anime.es.js';
+import './loading-screen.css';
 
-// const createDelayPromise = (timeout) => {
-//   var timeoutResolver = null;
+var loadingContainer = document.querySelector('.loading-container');
+loadingContainer.style.display = 'inline-table';
 
-//   window.setTimeout(function () {
-//     timeoutResolver();
-//   }, timeout);
+// Wrap every letter in a span
+var textWrapper = document.querySelector('.ml2');
+textWrapper.innerHTML = textWrapper.textContent.replace(/\S/g, "<span class='letter'>$&</span>");
 
-//   return new Promise((resolve) => {
-//     timeoutResolver = resolve;
-//   });
-// }
+anime.timeline({ loop: true })
+    .add({
+        targets: '.ml2 .letter',
+        scale: [4, 1],
+        opacity: [0, 1],
+        translateZ: 0,
+        easing: "easeOutExpo",
+        duration: 950,
+        delay: (el, i) => 70 * i
+    }).add({
+        targets: '.ml2',
+        opacity: 0,
+        duration: 1000,
+        easing: "easeOutExpo",
+        delay: 1000
+    });
 
-
-// const AppLazy = lazy(
-//   () => {
-//     const delayPromise = createDelayPromise(500);
-//     const appLoadingPromise = import('./app');
-
-//     // return appLoadingPromise
-//     return Promise.all([appLoadingPromise, delayPromise]).then((promises) => {
-//       return promises[0];
-//     });
-//   }
-// );
-
-// const App = () => {
-//   return (
-//     <Suspense fallback={
-//       <div style={{ margin: "10px", width: "100%", fontSize: "1.5em" }}>
-//         Loading...
-//       </div>
-//     }>
-//       <AppLazy />
-//     </Suspense>
-//   )
-// }
-
-var loadingContainer = document.querySelector(".loading-container");
-
-if (loadingContainer) {
-  loadingContainer.parentNode.removeChild(loadingContainer);
-}
-
-ReactDOM.render(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>,
-  document.getElementById('app-react-root')
-)
+import('./index2');
